Update .bash_load 10_jenv: prevent multiple hook loads
This can happen when dropping into subshells and, aside from being
unnecessary, will mess up shims on tools like sdkman. In general it is
likely expected that jenv will look after the java version and sdkman
will look after all other tools if both are installed.

It is still the case that when both are installed the sdkman init should
be run after the jenv init